Add querier test
