+++ modules/auth/test_utils.go
@@ -1,0 +1,151 @@
+// nolint
+package auth
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/irisnet/irishub/codec"
+	"github.com/irisnet/irishub/modules/params"
+	"github.com/irisnet/irishub/store"
+	sdk "github.com/irisnet/irishub/types"
+	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/crypto"
+	"github.com/tendermint/tendermint/crypto/ed25519"
+	"github.com/tendermint/tendermint/crypto/secp256k1"
+	dbm "github.com/tendermint/tendermint/libs/db"
+	"github.com/tendermint/tendermint/libs/log"
+)
+
+type testInput struct {
+	cdc *codec.Codec
+	ctx sdk.Context
+	ak  AccountKeeper
+	fck FeeKeeper
+}
+
+func setupTestInput() testInput {
+	db := dbm.NewMemDB()
+
+	cdc := codec.New()
+	RegisterBaseAccount(cdc)
+
+	authCapKey := sdk.NewKVStoreKey("authCapKey")
+	fckCapKey := sdk.NewKVStoreKey("fckCapKey")
+	keyParams := sdk.NewKVStoreKey("params")
+	tkeyParams := sdk.NewTransientStoreKey("transient_params")
+
+	ms := store.NewCommitMultiStore(db)
+	ms.MountStoreWithDB(authCapKey, sdk.StoreTypeIAVL, db)
+	ms.MountStoreWithDB(fckCapKey, sdk.StoreTypeIAVL, db)
+	ms.MountStoreWithDB(keyParams, sdk.StoreTypeIAVL, db)
+	ms.MountStoreWithDB(tkeyParams, sdk.StoreTypeTransient, db)
+	ms.LoadLatestVersion()
+
+	pk := params.NewKeeper(cdc, keyParams, tkeyParams)
+	ak := NewAccountKeeper(cdc, authCapKey, ProtoBaseAccount)
+	fck := NewFeeKeeper(cdc, fckCapKey, pk.Subspace(DefaultParamSpace))
+	ctx := sdk.NewContext(ms, abci.Header{ChainID: "test-chain-id"}, false, log.NewNopLogger())
+
+	return testInput{cdc: cdc, ctx: ctx, ak: ak, fck: fck}
+}
+
+func keyPubAddr() (crypto.PrivKey, crypto.PubKey, sdk.AccAddress) {
+	key := secp256k1.GenPrivKey()
+	pub := key.PubKey()
+	addr := sdk.AccAddress(pub.Address())
+	return key, pub, addr
+}
+
+func newTestMsg(addrs ...sdk.AccAddress) *sdk.TestMsg {
+	return sdk.NewTestMsg(addrs...)
+}
+
+func newStdFee() StdFee {
+	return NewStdFee(5000,
+		sdk.NewInt64Coin("atom", 150),
+	)
+}
+
+// coins to more than cover the fee
+func newCoins() sdk.Coins {
+	return sdk.Coins{
+		sdk.NewInt64Coin("atom", 10000000),
+	}
+}
+
+// generate a priv key and return it with its address
+func privAndAddr() (crypto.PrivKey, sdk.AccAddress) {
+	priv := ed25519.GenPrivKey()
+	addr := sdk.AccAddress(priv.PubKey().Address())
+	return priv, addr
+}
+
+// run the tx through the anteHandler and ensure its valid
+func checkValidTx(t *testing.T, anteHandler sdk.AnteHandler, ctx sdk.Context, tx sdk.Tx, simulate bool) {
+	_, result, abort := anteHandler(ctx, tx, simulate)
+	require.False(t, abort)
+	require.Equal(t, sdk.CodeOK, result.Code)
+	require.True(t, result.IsOK())
+}
+
+// run the tx through the anteHandler and ensure it fails with the given code
+func checkInvalidTx(t *testing.T, anteHandler sdk.AnteHandler, ctx sdk.Context, tx sdk.Tx, simulate bool, code sdk.CodeType) {
+	newCtx, result, abort := anteHandler(ctx, tx, simulate)
+	require.True(t, abort)
+	require.Equal(t, code, result.Code, fmt.Sprintf("Expected %v, got %v", code, result))
+	require.Equal(t, sdk.CodespaceRoot, result.Codespace)
+
+	if code == sdk.CodeOutOfGas {
+		stdTx, ok := tx.(StdTx)
+		require.True(t, ok, "tx must be in form auth.StdTx")
+		// GasWanted set correctly
+		require.Equal(t, stdTx.Fee.Gas, result.GasWanted, "Gas wanted not set correctly")
+		require.True(t, result.GasUsed > result.GasWanted, "GasUsed not greated than GasWanted")
+		// Check that context is set correctly
+		require.Equal(t, result.GasUsed, newCtx.GasMeter().GasConsumed(), "Context not updated correctly")
+	}
+}
+
+func newTestTx(ctx sdk.Context, msgs []sdk.Msg, privs []crypto.PrivKey, accNums []uint64, seqs []uint64, fee StdFee) sdk.Tx {
+	sigs := make([]StdSignature, len(privs))
+	for i, priv := range privs {
+		signBytes := StdSignBytes(ctx.ChainID(), accNums[i], seqs[i], fee, msgs, "")
+		sig, err := priv.Sign(signBytes)
+		if err != nil {
+			panic(err)
+		}
+		sigs[i] = StdSignature{PubKey: priv.PubKey(), Signature: sig, AccountNumber: accNums[i], Sequence: seqs[i]}
+	}
+	tx := NewStdTx(msgs, fee, sigs, "")
+	return tx
+}
+
+func newTestTxWithMemo(ctx sdk.Context, msgs []sdk.Msg, privs []crypto.PrivKey, accNums []uint64, seqs []uint64, fee StdFee, memo string) sdk.Tx {
+	sigs := make([]StdSignature, len(privs))
+	for i, priv := range privs {
+		signBytes := StdSignBytes(ctx.ChainID(), accNums[i], seqs[i], fee, msgs, memo)
+		sig, err := priv.Sign(signBytes)
+		if err != nil {
+			panic(err)
+		}
+		sigs[i] = StdSignature{PubKey: priv.PubKey(), Signature: sig, AccountNumber: accNums[i], Sequence: seqs[i]}
+	}
+	tx := NewStdTx(msgs, fee, sigs, memo)
+	return tx
+}
+
+// All signers sign over the same StdSignDoc. Should always create invalid signatures
+func newTestTxWithSignBytes(msgs []sdk.Msg, privs []crypto.PrivKey, accNums []uint64, seqs []uint64, fee StdFee, signBytes []byte, memo string) sdk.Tx {
+	sigs := make([]StdSignature, len(privs))
+	for i, priv := range privs {
+		sig, err := priv.Sign(signBytes)
+		if err != nil {
+			panic(err)
+		}
+		sigs[i] = StdSignature{PubKey: priv.PubKey(), Signature: sig, AccountNumber: accNums[i], Sequence: seqs[i]}
+	}
+	tx := NewStdTx(msgs, fee, sigs, memo)
+	return tx
+}
